+++ packages/react-components/src/pages/Button/ButtonSecondary.tsx
@@ -8,14 +8,14 @@ const iconPlaceholder = (
     fill="none"
     xmlns="http://www.w3.org/2000/svg"
   >
-    <g id="20px/Icon placeholder" clip-path="url(#clip0_717_6325)">
+    <g id="20px/Icon placeholder" clipPath="url(#clip0_717_6325)">
       <path
         id="Icon"
         d="M9.99996 18.3333C14.6023 18.3333 18.3333 14.6023 18.3333 9.99996C18.3333 5.39759 14.6023 1.66663 9.99996 1.66663C5.39759 1.66663 1.66663 5.39759 1.66663 9.99996C1.66663 14.6023 5.39759 18.3333 9.99996 18.3333Z"
         stroke="currentColor"
-        stroke-width="2"
-        stroke-linecap="round"
-        stroke-linejoin="round"
+        strokeWidth="2"
+        strokeLinecap="round"
+        strokeLinejoin="round"
       />
     </g>
     <defs>